more efficient loading

+++ test_savanna_conversion.py
@@ -46,7 +46,7 @@ if __name__ == "__main__":
     # config.cached_generation = False
     device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
     with torch.device(device):
-        m = StripedHyena(config).to(torch.float32)
+        m = StripedHyena(config)
 
     if args.checkpoint_path:
         state_dict = torch.load(args.checkpoint_path, map_location=device)
